[build] adding curl, fixing sprints

+++ capitalprojects_build/attributes/geom_from_id_bin_map.sql
@@ -5,5 +5,5 @@ SET geom = ST_Centroid(b.wkb_geometry),
     geomsource = 'footprint_script'
 FROM doitt_buildingfootprints b,
      dcp_id_bin_map c
-WHERE c.bin = b.bin::bigint AND
+WHERE c.bin::bigint = b.bin::bigint AND
       a.maprojid = c.maprojid;
